feat: refactor the drawer for editing the basic information of the experiment
