+++ chaosmeta-platform-frontend/src/pages/Space/AddExperiment/index.tsx
@@ -1,4 +1,21 @@
+// react
+import { useEffect, useState } from 'react';
+// UI 资产
+import { Button, Form, Modal, Space, Spin, message } from 'antd';
+import { PageContainer } from '@ant-design/pro-components';
+import { EditOutlined, ExclamationCircleFilled } from '@ant-design/icons';
+import ExperimentInfoEditDrawer from './components/ExperimentInfoEditDrawer';
 import ShowText from '@/components/ShowText';
+import ArrangeContent from './ArrangeContent';
+// 辅助函数
+import { history, useIntl, useModel, useRequest } from '@umijs/max';
+import NiceModal from '@ebay/nice-modal-react';
+import {
+  arrangeDataOriginTranstion,
+  arrangeDataResultTranstion,
+} from '@/utils/format';
+import { renderScheduleType, renderTags } from '@/utils/renderItem';
+// 网络请求
 import {
   createExperiment,
   deleteExperiment,
@@ -6,49 +23,40 @@ import {
   updateExperiment,
 } from '@/services/chaosmeta/ExperimentController';
 import { querySpaceUserPermission } from '@/services/chaosmeta/SpaceController';
-import {
-  arrangeDataOriginTranstion,
-  arrangeDataResultTranstion,
-} from '@/utils/format';
-import { renderScheduleType, renderTags } from '@/utils/renderItem';
-import { EditOutlined, ExclamationCircleFilled } from '@ant-design/icons';
-import { PageContainer } from '@ant-design/pro-components';
-import { history, useIntl, useModel, useRequest } from '@umijs/max';
-import { Button, Form, Modal, Space, Spin, message } from 'antd';
-import { useEffect, useState } from 'react';
-import ArrangeContent from './ArrangeContent';
-import InfoDrawer from './components/InfoDrawer';
+// 样式文件
 import { Container } from './style';
 
-const AddExperiment = () => {
+export default () => {
+  // 表单实例
   const [form] = Form.useForm();
   // 用户权限
   const { setSpacePermission, spacePermission } = useModel('global');
   // 编排的数据
   const [arrangeList, setArrangeList] = useState<any>([]);
-  // 编辑基本信息抽屉
-  const [infoDrawerOpen, setInfoDrawerOpen] = useState(false);
+  // 实验基本信息
   const [baseInfo, setBaseInfo] = useState<any>({});
+  /** 国际化 */
   const intl = useIntl();
+  /** 当前页面对应实验 id */
+  const experimentId = history?.location?.query?.experimentId;
 
-  /**
-   * 获取实验详情
-   */
+  /** 获取实验详情 */
   const getExperimentDetail = useRequest(queryExperimentDetail, {
     manual: true,
     formatResult: (res) => res,
     onSuccess: (res) => {
       if (res?.code === 200) {
         const experiments = res?.data?.experiments;
-        // 已经保存过的信息，完善度设为true，已完善
+        // 已经保存过的信息，完善度设为 true，已完善
         const newList = experiments?.workflow_nodes?.map((item: any) => {
-          // 将动态表单args_value的值处理为form可以使用的
+          // 将动态表单 args_value 的值处理为 form 可以使用的
           const newArgs: any = {};
           item?.args_value?.forEach((arg: any) => {
             newArgs[arg?.args_id] = arg?.value;
           });
           return { ...item, nodeInfoState: true, args_value: newArgs };
         });
+        // 保存实验基本信息
         form.setFieldsValue(experiments);
         setBaseInfo(experiments);
         setArrangeList(arrangeDataOriginTranstion(newList || []));
@@ -56,9 +64,7 @@ const AddExperiment = () => {
     },
   });
 
-  /**
-   * 编辑更新实验
-   */
+  /** 编辑更新实验 */
   const editExperiment = useRequest(updateExperiment, {
     manual: true,
     formatResult: (res) => res,
@@ -70,9 +76,7 @@ const AddExperiment = () => {
     },
   });
 
-  /**
-   * 创建实验
-   */
+  /** 创建实验 */
   const handleCreateExperiment = useRequest(createExperiment, {
     manual: true,
     formatResult: (res) => res,
@@ -84,9 +88,7 @@ const AddExperiment = () => {
     },
   });
 
-  /**
-   * 根据成员名称和空间id获取成员空间内权限信息
-   */
+  /** 根据成员名称和空间 id 获取成员空间内权限信息 */
   const getUserSpaceAuth = useRequest(querySpaceUserPermission, {
     manual: true,
     formatResult: (res) => res,
@@ -98,46 +100,11 @@ const AddExperiment = () => {
     },
   });
 
-  // 标题渲染
-  const renderTitle = () => {
-    return (
-      <Form form={form}>
-        {/* 用于保存内部弹窗中所填写的集群信息 */}
-        <Form.Item name='cluster_id' noStyle hidden />
-        <Space>
-          <Form.Item name={'name'}>
-            <ShowText ellipsis />
-          </Form.Item>
-          <>
-            {spacePermission === 1 ? (
-              <EditOutlined
-                className="edit"
-                style={{ color: '#1890FF' }}
-                onClick={() => {
-                  setInfoDrawerOpen(true);
-                }}
-              />
-            ) : (
-              <a
-                onClick={() => {
-                  setInfoDrawerOpen(true);
-                }}
-              >
-                {intl.formatMessage({ id: 'check' })}
-              </a>
-            )}
-          </>
-        </Space>
-        <Form.Item>{renderTags(baseInfo?.labels)}</Form.Item>
-      </Form>
-    );
-  };
+  /** 处理当前实验信息以供编辑或新增 */
+  const handleFormatOriginData = async () => {
+    try {
+      const values = await form.validateFields();
 
-  /**
-   * 提交实验信息
-   */
-  const handleSubmit = () => {
-    form.validateFields().then((values) => {
       const arrangeResult = arrangeDataResultTranstion(arrangeList);
       if (!baseInfo?.name || !baseInfo?.schedule_type) {
         message.info(intl.formatMessage({ id: 'addExperiment.basic.tip' }));
@@ -206,14 +173,24 @@ const AddExperiment = () => {
           flow_range,
         };
       });
-      const params = {
+
+      return {
         ...values,
         labels: newLabels,
         schedule_rule: baseInfo?.schedule_rule,
         namespace_id: Number(history?.location?.query?.spaceId),
         workflow_nodes: newList,
       };
-      const experimentId = history?.location?.query?.experimentId;
+    }
+    catch {
+      return undefined;
+    }
+  }
+
+  /** 提交实验信息 */
+  const handleSubmit = () => {
+    handleFormatOriginData().then((params) => {
+      // 根据新增还是编辑调用不同的接口
       if (experimentId) {
         editExperiment?.run({ ...params, uuid: experimentId });
       } else {
@@ -222,9 +199,7 @@ const AddExperiment = () => {
     });
   };
 
-  /**
-   * 删除实验接口
-   */
+  /** 删除实验接口 */
   const handleDeleteExperiment = useRequest(deleteExperiment, {
     manual: true,
     formatResult: (res) => res,
@@ -236,9 +211,7 @@ const AddExperiment = () => {
     },
   });
 
-  /**
-   * 确认删除实验
-   */
+  /** 确认删除实验 */
   const handleDeleteConfirm = () => {
     const uuid = history?.location?.query?.experimentId as string;
     if (uuid) {
@@ -253,65 +226,10 @@ const AddExperiment = () => {
     }
   };
 
-  const headerExtra = () => {
-    return (
-      <Form form={form}>
-        <div className="header-extra">
-          <div>
-            <Form.Item
-              name={'schedule_type'}
-              label={intl.formatMessage({ id: 'triggerMode' })}
-            >
-              {renderScheduleType(baseInfo)}
-            </Form.Item>
-            <Form.Item
-              name={'description'}
-              label={intl.formatMessage({ id: 'description' })}
-            >
-              <ShowText />
-            </Form.Item>
-          </div>
-          {spacePermission === 1 && (
-            <Space>
-              <Button
-                ghost
-                danger
-                onClick={() => {
-                  handleDeleteConfirm();
-                }}
-              >
-                {intl.formatMessage({ id: 'delete' })}
-              </Button>
-              <Button
-                ghost
-                type="primary"
-                loading={handleCreateExperiment?.loading}
-                onClick={() => {
-                  handleSubmit();
-                }}
-              >
-                {intl.formatMessage({ id: 'finish' })}
-              </Button>
-            </Space>
-          )}
-        </div>
-      </Form>
-    );
-  };
-
-  /**
-   * 更新基础信息
-   * @param values
-   */
-  const handleConfirm = (values: any) => {
-    form.setFieldsValue(values);
-    setBaseInfo({ ...baseInfo, ...values });
-  };
-
   useEffect(() => {
     const { experimentId, spaceId } = history?.location?.query || {};
     const sessionSpaceId = sessionStorage.getItem('spaceId');
-    // 地址栏中存在空间id，需要将空间列表选项更新，并保存当前id
+    // 地址栏中存在空间 id，需要将空间列表选项更新，并保存当前id
     if (spaceId || sessionSpaceId) {
       if (!spaceId) {
         history.push({
@@ -337,33 +255,113 @@ const AddExperiment = () => {
   return (
     <Container>
       <Spin spinning={getExperimentDetail.loading}>
-        <PageContainer
-          header={{
-            title: renderTitle(),
-            onBack: () => {
-              history.push('/space/experiment');
-            },
-            extra: headerExtra(),
-          }}
-        >
-          <ArrangeContent
-            arrangeList={arrangeList}
-            setArrangeList={setArrangeList}
-            disabled={spacePermission !== 1}
-          />
-          {infoDrawerOpen && (
-            <InfoDrawer
-              open={infoDrawerOpen}
-              setOpen={setInfoDrawerOpen}
-              spacePermission={spacePermission}
-              handleConfirm={handleConfirm}
-              baseInfo={baseInfo}
+        <NiceModal.Provider>
+          <PageContainer
+            header={{
+              title: (
+                <Form form={form}>
+                  {/* 用于保存内部弹窗中所填写的集群信息 */}
+                  <Form.Item name='cluster_id' noStyle hidden />
+                  <Space>
+                    <Form.Item name={'name'}>
+                      <ShowText ellipsis />
+                    </Form.Item>
+                    <EditOutlined
+                      hidden={spacePermission !== 1}
+                      className="edit"
+                      style={{ color: '#1890FF' }}
+                      onClick={async () => {
+                        const formData = await NiceModal.show(ExperimentInfoEditDrawer, { baseInfo, spacePermission });
+
+                        if (formData instanceof Object) {
+                          // 编辑时直接调用接口更新并刷新当前页面
+                          if (experimentId) {
+                            try {
+                              const data = await handleFormatOriginData();
+
+                              await editExperiment?.run({ ...data, ...formData, uuid: experimentId });
+                              window.location.reload();
+                            }
+                            catch {
+                              // 接口报错则暂时记录当前信息
+                              setBaseInfo({ ...baseInfo, ...formData });
+                              form.setFieldsValue(formData);
+                            }
+                          } else {
+                            setBaseInfo({ ...baseInfo, ...formData });
+                            form.setFieldsValue(formData);
+                          }
+                        }
+                      }}
+                    />
+                    <a
+                      hidden={spacePermission === 1}
+                      onClick={() => {
+                        NiceModal.show(ExperimentInfoEditDrawer, { baseInfo, spacePermission });
+                      }}
+                    >
+                      {intl.formatMessage({ id: 'check' })}
+                    </a>
+                  </Space>
+                  <Form.Item>{renderTags(baseInfo?.labels)}</Form.Item>
+                </Form>
+              ),
+              onBack: () => {
+                history.push('/space/experiment');
+              },
+              extra: (
+                <Form form={form}>
+                  <div className="header-extra">
+                    <div>
+                      <Form.Item
+                        name={'schedule_type'}
+                        label={intl.formatMessage({ id: 'triggerMode' })}
+                      >
+                        {renderScheduleType(baseInfo)}
+                      </Form.Item>
+                      <Form.Item
+                        name={'description'}
+                        label={intl.formatMessage({ id: 'description' })}
+                      >
+                        <ShowText />
+                      </Form.Item>
+                    </div>
+                    {spacePermission === 1 && (
+                      <Space>
+                        <Button
+                          ghost
+                          danger
+                          onClick={() => {
+                            handleDeleteConfirm();
+                          }}
+                        >
+                          {intl.formatMessage({ id: 'delete' })}
+                        </Button>
+                        <Button
+                          ghost
+                          type="primary"
+                          loading={handleCreateExperiment?.loading}
+                          onClick={() => {
+                            handleSubmit();
+                          }}
+                        >
+                          {intl.formatMessage({ id: 'finish' })}
+                        </Button>
+                      </Space>
+                    )}
+                  </div>
+                </Form>
+              ),
+            }}
+          >
+            <ArrangeContent
+              arrangeList={arrangeList}
+              setArrangeList={setArrangeList}
+              disabled={spacePermission !== 1}
             />
-          )}
-        </PageContainer>
+          </PageContainer>
+        </NiceModal.Provider>
       </Spin>
     </Container>
   );
 };
-
-export default AddExperiment;
